+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: Filecoin FEVM token support
 - fixed: Zano deeplink parsing bug when scanning for unknown currency.
 
 ## 4.53.0 (2025-07-07)